Moving some parts to match existing dir structure.
