+++ tripal_chado/src/Form/ChadoTermMappingDeleteForm.php
@@ -6,7 +6,7 @@ use Drupal\Core\Url;
 use Drupal\Core\Form\FormStateInterface;
 
 /**
- * Builds the form to delete an Example.
+ * Builds the form to delete an Chado Term Mapping Config Entity.
  */
 
 class ChadoTermMappingDeleteForm extends EntityConfirmFormBase {
